Monastery-Preservation/python-service/requirements.txt
flask
opencv-python
scikit-image
numpy
Pillow
python-dotenv
